+++ spine-cpp/spine-cpp/include/spine/BoundingBoxAttachment.h
@@ -45,6 +45,7 @@ namespace spine {
 		Color &getColor();
 
 		virtual Attachment *copy();
+
 	private:
 		Color _color;
 	};