+++ spine-cpp/spine-cpp/src/spine/PathAttachment.cpp
@@ -37,7 +37,8 @@ using namespace spine;
 
 RTTI_IMPL(PathAttachment, VertexAttachment)
 
-PathAttachment::PathAttachment(const String &name) : VertexAttachment(name), _closed(false), _constantSpeed(false), _color() {
+PathAttachment::PathAttachment(const String &name) : VertexAttachment(name), _closed(false), _constantSpeed(false),
+													 _color() {
 }
 
 Vector<float> &PathAttachment::getLengths() {